move triggers and left over operators and sensors into standard provider
