+++ providers/src/airflow/providers/standard/operators/email.py
