+++ tests/operators/test_trigger_dagrun.py
@@ -32,8 +32,8 @@ from airflow.models.log import Log
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.models.taskinstance import TaskInstance
 from airflow.operators.trigger_dagrun import TriggerDagRunOperator
+from airflow.providers.standard.triggers.external_task import DagStateTrigger
 from airflow.settings import TracebackSessionForTests
-from airflow.triggers.external_task import DagStateTrigger
 from airflow.utils import timezone
 from airflow.utils.session import create_session
 from airflow.utils.state import DagRunState, State, TaskInstanceState
