+++ tests/system/core/example_external_task_parent_deferrable.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 from airflow import DAG
 from airflow.operators.empty import EmptyOperator
 from airflow.operators.trigger_dagrun import TriggerDagRunOperator
-from airflow.sensors.external_task import ExternalTaskSensor
+from airflow.providers.standard.sensors.external_task import ExternalTaskSensor
 from airflow.utils.timezone import datetime
 
 with DAG(
